# Optimize `DocumentKey` comparisons