--- a/Firestore/core/src/firebase/firestore/model/base_path.h
+++ b/Firestore/core/src/firebase/firestore/model/base_path.h
@@ -152,6 +152,7 @@ class BasePath {
 
   util::ComparisonResult CompareTo(const T& rhs) const {
     return util::Compare(segments_, rhs.segments_);
+    //return util::CompareContainer(segments_, rhs.segments_);
   }
 
   size_t Hash() const {