--- a/Firestore/core/src/firebase/firestore/model/base_path.h
+++ b/Firestore/core/src/firebase/firestore/model/base_path.h
@@ -151,8 +151,12 @@ class BasePath {
   }
 
   util::ComparisonResult CompareTo(const T& rhs) const {
-    return util::Compare(segments_, rhs.segments_);
-    //return util::CompareContainer(segments_, rhs.segments_);
+    //return util::Compare(segments_, rhs.segments_);
+    return util::CompareContainer(segments_, rhs.segments_);
+  }
+
+  bool IsEqual(const T& rhs) const {
+    return segments_ == rhs.segments_;
   }
 
   size_t Hash() const {