--- a/Firestore/core/src/firebase/firestore/model/base_path.h
+++ b/Firestore/core/src/firebase/firestore/model/base_path.h
@@ -151,7 +151,6 @@ class BasePath {
   }
 
   util::ComparisonResult CompareTo(const T& rhs) const {
-    //return util::Compare(segments_, rhs.segments_);
     return util::CompareContainer(segments_, rhs.segments_);
   }