--- a/Firestore/core/src/firebase/firestore/model/document_key.h
+++ b/Firestore/core/src/firebase/firestore/model/document_key.h
@@ -35,7 +35,7 @@ namespace model {
 /**
  * DocumentKey represents the location of a document in the Firestore database.
  */
-class DocumentKey : public util::Comparable<DocumentKey> {
+class DocumentKey : public util::ComparableWithEqual<DocumentKey> {
  public:
   /** Creates a "blank" document key not associated with any document. */
   DocumentKey() : path_{std::make_shared<ResourcePath>()} {
@@ -69,6 +69,8 @@ class DocumentKey : public util::Comparable<DocumentKey> {
   }
 
   util::ComparisonResult CompareTo(const DocumentKey& other) const;
+
+  bool IsEqual(const DocumentKey& other) const;
 
   size_t Hash() const {
     return util::Hash(ToString());