--- a/Firestore/core/src/firebase/firestore/model/document_key.h
+++ b/Firestore/core/src/firebase/firestore/model/document_key.h
@@ -35,7 +35,7 @@ namespace model {
 /**
  * DocumentKey represents the location of a document in the Firestore database.
  */
-class DocumentKey : public util::ComparableWithEqual<DocumentKey> {
+class DocumentKey : public util::InequalityComparable<DocumentKey> {
  public:
   /** Creates a "blank" document key not associated with any document. */
   DocumentKey() : path_{std::make_shared<ResourcePath>()} {
@@ -70,7 +70,9 @@ class DocumentKey : public util::ComparableWithEqual<DocumentKey> {
 
   util::ComparisonResult CompareTo(const DocumentKey& other) const;
 
-  bool IsEqualTo(const DocumentKey& other) const;
+  friend bool operator==(const DocumentKey& lhs, const DocumentKey& rhs) {
+    return lhs.path() == rhs.path();
+  }
 
   size_t Hash() const {
     return util::Hash(ToString());