@var-const 2019-10-19
@wilhuff Do you think this warrants a changelog entry? On one hand, up to 10% CPU time reduction is non-negligible; on the other hand, wall time is dominated by network, so it's unlikely to be very noticeable to the end users, and I don't want to "oversell" the effects of this change.