@wilhuff 2019-10-20
Wall time is what matters and since the change isn’t significant I’d lean toward skipping the changelog entry.